backend/requirements.txt
AI GENERATED FILE

fastapi
uvicorn
pydantic
python-multipart
chromadb
httpx
beautifulsoup4
pytest
python-dotenv
